refactor(user): 유저 검색 api를 user 도메인으로 변경
